requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
Flask>=2.0.0
Flask-Cors>=3.0.10
groq
python-dotenv>=1.0.0
